+++ src/IM/Utils.java
@@ -26,8 +26,8 @@ public class Utils {
         }
         return url.substring(index + 1);
     }
-    public static int[][] getMatrix(BufferedImage bufferedImage) {
-        int[][] matrix = new int[bufferedImage.getWidth()][bufferedImage.getHeight()];
+    public static double[][] getMatrix(BufferedImage bufferedImage) {
+        double[][] matrix = new double[bufferedImage.getWidth()][bufferedImage.getHeight()];
 
         for (int i = 0; i < bufferedImage.getWidth(); i++) {
             for (int j = 0; j < bufferedImage.getHeight(); j++) {
@@ -38,7 +38,7 @@ public class Utils {
         return matrix;
     }
 
-    public static BufferedImage setImage(int[][] matrix) {
+    public static BufferedImage setImage(double[][] matrix) {
         BufferedImage image = new BufferedImage(matrix.length, matrix[0].length, BufferedImage.TYPE_4BYTE_ABGR);
         for (int i = 0; i < image.getWidth(); i++) {
             for (int j = 0; j < image.getHeight(); j++) {
@@ -47,7 +47,7 @@ public class Utils {
                 } else if (matrix[i][j] < 0) {
                     matrix[i][j] = 0;
                 }
-                image.setRGB(i, j, ((0xFF) << 24) | (matrix[i][j] << 16) | (matrix[i][j] << 8) | matrix[i][j]);
+                image.setRGB(i, j, ((0xFF) << 24) | ((int)matrix[i][j] << 16) | ((int)matrix[i][j] << 8) | (int)matrix[i][j]);
             }
         }
         return image;
